Receiver: modify decode process to automatically determine the bit periods
Instead of assuming that a 1-bit is 3x the duration of a 0-bit:
- The 1-bit is at least 2x the duration of the 0-bit
- The bit duration is allowed to vary by ±50%
- The post-sync duration is allowed to vary relative to the pre-sync duration

The first 6 incoming bits are sampled and used for timing of the rest of the
bits, as long as a 1-bit has been seen. Otherwise up to 16 bits in total are
sampled in an attempt to find the duration of a 1-bit.